Adds Data Overlay Filter
