+++ include/DataOverlayFilter.h
@@ -1,0 +1,12 @@
+#include "VisionFilter.h"
+
+namespace Icarus
+{
+  class DataOverlayFilter : public VisionFilter
+  {
+    public:
+      DataOverlayFilter();
+      virtual ~DataOverlayFilter();
+      void Filter(ImageData* data);
+  };
+}
